Nav Bar
This was a first draft of trying to duplicate the nav-bar. I ran into a few issues, the biggest ones involving the icons. The icons I used are not the same and are not all transparent. I would pay to use font-awesome but, as this is project is just for practice, I didn't feel it was necessary.

I also ran into some trouble with the profile icon inside the button on the right. I have commented it out for now and will return to it later. I need an alternative to flex-containers as this isn't compatable with a button.

I'll leave the nav as it is for now so I can get started with the rest of the page.