Video Area
In this commit I've adjusted the size of the video being displayed. I also changed the background color of the body, so the contrast between the white nav bar and the darker white color of the body can be seen.

I've seperated the video info section into two flex containers, each one will display the video info and options (like, share etc) respectively. I've changed the background colors temporarily so this can be seen more clearly. I'll continue to work in these containers in the next commit.