Advert
Here I've duplicated the layout of the advert section. I split the div into 2 more divs (with a flex-direction of column) and positioned elements within their repsective flex containers.

This was the last section of the page. I've found this to be a useful exercise to pracice Grid and Flexbox. I now feel more confident in positioning elements where I want them on a page. I'll continue to practice and keep trying to improve my skills.